+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/rest/CreateInvite.kt
@@ -1,0 +1,12 @@
+package com.jessecorbett.diskord.api.rest
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CreateInvite(
+        @SerialName("max_age") val expiresInSeconds: Int,
+        @SerialName("max_uses") val maxUses: Int,
+        @SerialName("temporary") val temporaryMembership: Boolean,
+        @SerialName("unique") val doNotAttemptReuse: Boolean
+)