+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/websocket/events/ChannelPinUpdate.kt
@@ -1,0 +1,10 @@
+package com.jessecorbett.diskord.api.websocket.events
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ChannelPinUpdate(
+        @SerialName("channel_id") val channelId: String,
+        @SerialName("last_pin_timestamp") val lastPinAt: String?
+)