+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/websocket/events/MessageReactionRemoveAll.kt
@@ -1,0 +1,10 @@
+package com.jessecorbett.diskord.api.websocket.events
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MessageReactionRemoveAll(
+        @SerialName("channel_id") val channelId: String,
+        @SerialName("message_id") val messageId: String
+)