+++ src/jvmMain/kotlin/com/jessecorbett/diskord/api/rest/CreateWebhook.kt
@@ -1,8 +1,0 @@
-package com.jessecorbett.diskord.api.rest
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class CreateWebhook(
-        @JsonProperty("name") val name: String,
-        @JsonProperty("avatar") val base64AvatarData: String? = null
-)