+++ src/jvmMain/kotlin/com/jessecorbett/diskord/api/rest/PatchEmoji.kt
@@ -1,8 +1,0 @@
-package com.jessecorbett.diskord.api.rest
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class PatchEmoji(
-        @JsonProperty("name") val name: String,
-        @JsonProperty("roles") val roles: List<String>
-)